+++ apps/web-svelte/src/routes/galleries/[id]/+page.svelte
@@ -331,6 +331,7 @@
           cards={imageChildren}
           prefsKey={`gallery-${gallery?.id}-images`}
           selectable={false}
+          initialMediaWall
           onCardActivate={openImageLightbox}
           emptyTitle="No images"
           emptyMessage="This gallery has no images."